Moves DAG.fit method to DiscreteBayesianNetwork.fit
